Move tests around to mimic sources folder setup
