+++ Tests/SuperwallKitTests/Session Events/SessionEventsDelegateMock.swift
