+++ Tests/SuperwallKitTests/Session Events/Trigger Session Manager/MockSkProduct.swift
